requirements.txt
python-telegram-bot
speedtest-cli
schedule
matplotlib
pandas
ping3